refactor(fmir): Simple manual organization.
