+++ compiler/cx-safe-analyzer/src/mir_conversion/expression.rs
@@ -1,409 +1,16 @@
 use cx_mir::mir::{
     expression::{
-        MIRBinOp, MIRCoercion, MIRExpression, MIRExpressionKind, MIRFloatBinOp, MIRIntegerBinOp,
-        MIRPtrBinOp, MIRPtrDiffBinOp, MIRUnOp,
+        MIRExpression, MIRExpressionKind, MIRUnOp,
     },
-    types::{MIRIntegerType, MIRType, MIRTypeKind},
+    types::{MIRType, MIRTypeKind},
 };
-use cx_safe_ir::ast::{
-    CVMOperation, FMIRBinaryIntrinsic, FMIRCastIntrinsic, FMIRFloatBinaryIntrinsicOp,
-    FMIRIntegerBinaryIntrinsicOp, FMIRIntrinsicFunction, FMIRIntrinsicKind, FMIRNode,
-    FMIRNodeBody, FMIRPointerBinaryIntrinsicOp, FMIRPointerDiffBinaryIntrinsicOp, FMIRSourceRange,
-    FMIRType, FMIRUnaryIntrinsic, FRc, MemoryLocation,
-};
+use cx_safe_ir::{ast::{
+    CVMOperation, FMIRNode, FMIRNodeBody, FMIRSourceRange, FMIRType, FRc, MemoryLocation,
+}, intrinsic::FMIRIntrinsicKind};
+use crate::mir_conversion::factories::*;
 use cx_util::{CXError, CXResult, identifier::CXIdent};
 
 use crate::mir_conversion::environment::FMIREnvironment;
-
-fn monad_unit(operation: CVMOperation) -> FMIRType {
-    FMIRType::CMonad {
-        inner: Box::new(FMIRType::pure(MIRType::unit())),
-        operation,
-    }
-}
-
-fn pointer_alias(name: &CXIdent, pointer_type: MIRType) -> FMIRNode {
-    FMIRNode {
-        source_range: None,
-        body: FMIRNodeBody::VariableAlias {
-            name: name.as_string(),
-        },
-        _type: FMIRType::pure(pointer_type),
-    }
-}
-
-fn intrinsic_alias(intrinsic: FMIRIntrinsicKind) -> FMIRNode {
-    FMIRNode {
-        source_range: None,
-        body: FMIRNodeBody::IntrinsicFunction(FMIRIntrinsicFunction { kind: intrinsic }),
-        _type: FMIRType::pure(MIRType::internal_function()),
-    }
-}
-
-fn then_node(first: FMIRNode, second: FMIRNode) -> FMIRNode {
-    let combined = first._type.union(&second._type);
-    FMIRNode {
-        source_range: None,
-        _type: combined.apply(second._type.inner_type().clone()),
-        body: FMIRNodeBody::Then {
-            first: FRc::new(first),
-            second: FRc::new(second),
-        },
-    }
-}
-
-fn bind_node(monad: FMIRNode, capture: CXIdent, function: FMIRNode) -> FMIRNode {
-    let combined = monad._type.union(&function._type);
-    FMIRNode {
-        source_range: None,
-        _type: combined.apply(function._type.inner_type().clone()),
-        body: FMIRNodeBody::Bind {
-            monad: FRc::new(monad),
-            capture,
-            function: FRc::new(function),
-        },
-    }
-}
-
-fn chain_statements(statements: Vec<FMIRNode>) -> FMIRNode {
-    let mut iter = statements.into_iter();
-    let Some(first) = iter.next() else {
-        return FMIRNode::unit();
-    };
-
-    iter.fold(first, then_node)
-}
-
-fn unary_op_intrinsic(op: &MIRUnOp) -> Option<FMIRUnaryIntrinsic> {
-    Some(match op {
-        MIRUnOp::NEG => FMIRUnaryIntrinsic::Neg,
-        MIRUnOp::INEG => FMIRUnaryIntrinsic::INeg,
-        MIRUnOp::FNEG => FMIRUnaryIntrinsic::FNeg,
-        MIRUnOp::BNOT => FMIRUnaryIntrinsic::BNot,
-        MIRUnOp::LNOT => FMIRUnaryIntrinsic::LNot,
-        MIRUnOp::PreIncrement(_) | MIRUnOp::PostIncrement(_) => return None,
-    })
-}
-
-fn int_binop_intrinsic(op: &MIRIntegerBinOp) -> FMIRIntegerBinaryIntrinsicOp {
-    match op {
-        MIRIntegerBinOp::ADD => FMIRIntegerBinaryIntrinsicOp::Add,
-        MIRIntegerBinOp::SUB => FMIRIntegerBinaryIntrinsicOp::Sub,
-        MIRIntegerBinOp::MUL => FMIRIntegerBinaryIntrinsicOp::Mul,
-        MIRIntegerBinOp::DIV => FMIRIntegerBinaryIntrinsicOp::Div,
-        MIRIntegerBinOp::MOD => FMIRIntegerBinaryIntrinsicOp::Mod,
-        MIRIntegerBinOp::IMUL => FMIRIntegerBinaryIntrinsicOp::IMul,
-        MIRIntegerBinOp::IDIV => FMIRIntegerBinaryIntrinsicOp::IDiv,
-        MIRIntegerBinOp::IMOD => FMIRIntegerBinaryIntrinsicOp::IMod,
-        MIRIntegerBinOp::EQ => FMIRIntegerBinaryIntrinsicOp::Eq,
-        MIRIntegerBinOp::NE => FMIRIntegerBinaryIntrinsicOp::Ne,
-        MIRIntegerBinOp::LT => FMIRIntegerBinaryIntrinsicOp::Lt,
-        MIRIntegerBinOp::LE => FMIRIntegerBinaryIntrinsicOp::Le,
-        MIRIntegerBinOp::GT => FMIRIntegerBinaryIntrinsicOp::Gt,
-        MIRIntegerBinOp::GE => FMIRIntegerBinaryIntrinsicOp::Ge,
-        MIRIntegerBinOp::ILT => FMIRIntegerBinaryIntrinsicOp::ILt,
-        MIRIntegerBinOp::ILE => FMIRIntegerBinaryIntrinsicOp::ILe,
-        MIRIntegerBinOp::IGT => FMIRIntegerBinaryIntrinsicOp::IGt,
-        MIRIntegerBinOp::IGE => FMIRIntegerBinaryIntrinsicOp::IGe,
-        MIRIntegerBinOp::LAND => FMIRIntegerBinaryIntrinsicOp::LAnd,
-        MIRIntegerBinOp::LOR => FMIRIntegerBinaryIntrinsicOp::LOr,
-        MIRIntegerBinOp::BAND => FMIRIntegerBinaryIntrinsicOp::BAnd,
-        MIRIntegerBinOp::BOR => FMIRIntegerBinaryIntrinsicOp::BOr,
-        MIRIntegerBinOp::BXOR => FMIRIntegerBinaryIntrinsicOp::BXor,
-    }
-}
-
-fn float_binop_intrinsic(op: &MIRFloatBinOp) -> FMIRFloatBinaryIntrinsicOp {
-    match op {
-        MIRFloatBinOp::FADD => FMIRFloatBinaryIntrinsicOp::Add,
-        MIRFloatBinOp::FSUB => FMIRFloatBinaryIntrinsicOp::Sub,
-        MIRFloatBinOp::FMUL => FMIRFloatBinaryIntrinsicOp::Mul,
-        MIRFloatBinOp::FDIV => FMIRFloatBinaryIntrinsicOp::Div,
-        MIRFloatBinOp::EQ => FMIRFloatBinaryIntrinsicOp::Eq,
-        MIRFloatBinOp::NEQ => FMIRFloatBinaryIntrinsicOp::Ne,
-        MIRFloatBinOp::FLT => FMIRFloatBinaryIntrinsicOp::Lt,
-        MIRFloatBinOp::FLE => FMIRFloatBinaryIntrinsicOp::Le,
-        MIRFloatBinOp::FGT => FMIRFloatBinaryIntrinsicOp::Gt,
-        MIRFloatBinOp::FGE => FMIRFloatBinaryIntrinsicOp::Ge,
-    }
-}
-
-fn ptrdiff_binop_intrinsic(op: &MIRPtrDiffBinOp) -> FMIRPointerDiffBinaryIntrinsicOp {
-    match op {
-        MIRPtrDiffBinOp::ADD => FMIRPointerDiffBinaryIntrinsicOp::Add,
-        MIRPtrDiffBinOp::SUB => FMIRPointerDiffBinaryIntrinsicOp::Sub,
-    }
-}
-
-fn ptr_binop_intrinsic(op: &MIRPtrBinOp) -> FMIRPointerBinaryIntrinsicOp {
-    match op {
-        MIRPtrBinOp::EQ => FMIRPointerBinaryIntrinsicOp::Eq,
-        MIRPtrBinOp::NE => FMIRPointerBinaryIntrinsicOp::Ne,
-        MIRPtrBinOp::LT => FMIRPointerBinaryIntrinsicOp::Lt,
-        MIRPtrBinOp::GT => FMIRPointerBinaryIntrinsicOp::Gt,
-        MIRPtrBinOp::LE => FMIRPointerBinaryIntrinsicOp::Le,
-        MIRPtrBinOp::GE => FMIRPointerBinaryIntrinsicOp::Ge,
-    }
-}
-
-fn binary_op_intrinsic(op: &MIRBinOp) -> FMIRBinaryIntrinsic {
-    match op {
-        MIRBinOp::Integer { itype, op } => FMIRBinaryIntrinsic::Integer {
-            bits: itype.bytes() * 8,
-            op: int_binop_intrinsic(op),
-        },
-        MIRBinOp::Float { ftype, op } => FMIRBinaryIntrinsic::Float {
-            bits: ftype.bytes() * 8,
-            op: float_binop_intrinsic(op),
-        },
-        MIRBinOp::PtrDiff { op, .. } => FMIRBinaryIntrinsic::PointerDiff {
-            op: ptrdiff_binop_intrinsic(op),
-        },
-        MIRBinOp::Pointer { op } => FMIRBinaryIntrinsic::Pointer {
-            op: ptr_binop_intrinsic(op),
-        },
-    }
-}
-
-fn coercion_intrinsic(coercion: &MIRCoercion) -> FMIRCastIntrinsic {
-    match coercion {
-        MIRCoercion::Integral { sextend, to_type } => FMIRCastIntrinsic::Integral {
-            sextend: *sextend,
-            to_bits: to_type.bytes() * 8,
-        },
-        MIRCoercion::FloatCast { to_type } => FMIRCastIntrinsic::FloatCast {
-            to_bits: to_type.bytes() * 8,
-        },
-        MIRCoercion::PtrToInt { to_type } => FMIRCastIntrinsic::PtrToInt {
-            to_bits: to_type.bytes() * 8,
-        },
-        MIRCoercion::IntToPtr { sextend } => FMIRCastIntrinsic::IntToPtr {
-            sextend: *sextend,
-        },
-        MIRCoercion::IntToFloat { to_type, sextend } => FMIRCastIntrinsic::IntToFloat {
-            to_bits: to_type.bytes() * 8,
-            sextend: *sextend,
-        },
-        MIRCoercion::FloatToInt { to_type, sextend } => FMIRCastIntrinsic::FloatToInt {
-            to_bits: to_type.bytes() * 8,
-            sextend: *sextend,
-        },
-        MIRCoercion::IntToBool => FMIRCastIntrinsic::IntToBool,
-        MIRCoercion::ReinterpretBits => FMIRCastIntrinsic::ReinterpretBits,
-    }
-}
-
-fn app1(intrinsic: FMIRIntrinsicKind, arg: FMIRNode, output_type: &MIRType) -> FMIRNode {
-    FMIRNode {
-        source_range: None,
-        _type: FMIRType::pure(output_type.clone()),
-        body: FMIRNodeBody::Application {
-            function: FRc::new(intrinsic_alias(intrinsic)),
-            argument: FRc::new(arg),
-        },
-    }
-}
-
-fn app2(intrinsic: FMIRIntrinsicKind, lhs: FMIRNode, rhs: FMIRNode, output_type: &MIRType) -> FMIRNode {
-    FMIRNode {
-        source_range: None,
-        _type: FMIRType::pure(output_type.clone()),
-        body: FMIRNodeBody::Application {
-            function: FRc::new(FMIRNode {
-                source_range: None,
-                _type: FMIRType::pure(MIRType::internal_function()),
-                body: FMIRNodeBody::Application {
-                    function: FRc::new(intrinsic_alias(intrinsic)),
-                    argument: FRc::new(lhs),
-                },
-            }),
-            argument: FRc::new(rhs),
-        },
-    }
-}
-
-fn source_variable_name(expr: &MIRExpression) -> Option<&CXIdent> {
-    match &expr.kind {
-        MIRExpressionKind::Variable(name) | MIRExpressionKind::ContractVariable { name, .. } => {
-            Some(name)
-        }
-        _ => None,
-    }
-}
-
-fn read_operation_for_expr(env: &FMIREnvironment, source: &MIRExpression) -> CVMOperation {
-    source_variable_name(source)
-        .and_then(|name| env.query_memory_location(name))
-        .map(|location| CVMOperation::Access {
-            reads: vec![location],
-            writes: vec![],
-        })
-        .unwrap_or(CVMOperation::Unsafe)
-}
-
-fn write_operation_for_expr(env: &FMIREnvironment, target: &MIRExpression) -> CVMOperation {
-    source_variable_name(target)
-        .and_then(|name| env.query_memory_location(name))
-        .map(|location| CVMOperation::Access {
-            reads: vec![],
-            writes: vec![location],
-        })
-        .unwrap_or(CVMOperation::Unsafe)
-}
-
-fn invalidate_known_value_for_expr(env: &mut FMIREnvironment, target: &MIRExpression) {
-    if let Some(name) = source_variable_name(target) {
-        env.set_known_value(name, None);
-    }
-}
-
-fn load_node(pointer: FMIRNode, value_type: &MIRType, operation: CVMOperation) -> FMIRNode {
-    let read_effect = FMIRType::CMonad {
-        inner: Box::new(FMIRType::pure(value_type.clone())),
-        operation,
-    };
-    let combined = pointer
-        ._type
-        .union(&read_effect)
-        .apply(FMIRType::pure(value_type.clone()));
-
-    FMIRNode {
-        source_range: None,
-        _type: combined,
-        body: FMIRNodeBody::Load {
-            pointer: FRc::new(pointer),
-        },
-    }
-}
-
-fn store_node(pointer: FMIRNode, value: FMIRNode, operation: CVMOperation) -> FMIRNode {
-    let write_effect = monad_unit(operation);
-    let combined = pointer
-        ._type
-        .union(&value._type)
-        .union(&write_effect)
-        .apply(FMIRType::pure(MIRType::unit()));
-
-    FMIRNode {
-        source_range: None,
-        _type: combined,
-        body: FMIRNodeBody::Store {
-            pointer: FRc::new(pointer),
-            value: FRc::new(value),
-        },
-    }
-}
-
-fn increment_amount_node(value: i64, mir_type: &MIRType) -> CXResult<FMIRNode> {
-    let MIRTypeKind::Integer { _type, signed } = &mir_type.kind else {
-        return CXError::create_result(format!(
-            "FMIR increment desugaring expected integer type, found '{}'",
-            mir_type
-        ));
-    };
-
-    Ok(FMIRNode {
-        source_range: None,
-        body: FMIRNodeBody::IntegerLiteral(value),
-        _type: FMIRType::pure(MIRType::from(MIRTypeKind::Integer {
-            _type: *_type,
-            signed: *signed,
-        })),
-    })
-}
-
-fn convert_increment(
-    env: &mut FMIREnvironment,
-    operand_expr: &MIRExpression,
-    amount: i8,
-    is_pre: bool,
-) -> CXResult<FMIRNode> {
-    let pointer_node = convert_expression(env, operand_expr)?;
-    let Some(value_type) = operand_expr._type.mem_ref_inner().cloned() else {
-        return CXError::create_result(format!(
-            "FMIR increment desugaring expected memory reference operand, found '{}'",
-            operand_expr._type
-        ));
-    };
-
-    let old_value_load = load_node(
-        pointer_node.clone(),
-        &value_type,
-        read_operation_for_expr(env, operand_expr),
-    );
-
-    let old_capture = CXIdent::from("__inc_old");
-    let old_alias = FMIRNode {
-        source_range: None,
-        body: FMIRNodeBody::VariableAlias {
-            name: old_capture.as_string(),
-        },
-        _type: FMIRType::pure(value_type.clone()),
-    };
-
-    let (add_intrinsic, delta_node) = match &value_type.kind {
-        MIRTypeKind::Integer { _type, .. } => (
-            FMIRIntrinsicKind::Binary(FMIRBinaryIntrinsic::Integer {
-                bits: _type.bytes() * 8,
-                op: FMIRIntegerBinaryIntrinsicOp::Add,
-            }),
-            increment_amount_node(i64::from(amount), &value_type)?,
-        ),
-        MIRTypeKind::PointerTo { .. } => {
-            let op = if amount < 0 {
-                FMIRPointerDiffBinaryIntrinsicOp::Sub
-            } else {
-                FMIRPointerDiffBinaryIntrinsicOp::Add
-            };
-            let delta_type = MIRType::from(MIRTypeKind::Integer {
-                _type: MIRIntegerType::I64,
-                signed: true,
-            });
-            (
-                FMIRIntrinsicKind::Binary(FMIRBinaryIntrinsic::PointerDiff { op }),
-                increment_amount_node(i64::from(amount).abs(), &delta_type)?,
-            )
-        }
-        _ => {
-            return CXError::create_result(format!(
-                "FMIR increment desugaring requires integer or pointer inner type, found '{}'",
-                value_type
-            ));
-        }
-    };
-
-    let new_value = app2(add_intrinsic, old_alias.clone(), delta_node, &value_type);
-
-    invalidate_known_value_for_expr(env, operand_expr);
-    let store = store_node(
-        pointer_node,
-        new_value.clone(),
-        write_operation_for_expr(env, operand_expr),
-    );
-
-    let result_value = if is_pre { new_value } else { old_alias };
-    Ok(bind_node(
-        old_value_load,
-        old_capture,
-        then_node(store, result_value),
-    ))
-}
-
-fn unsupported_expression_error(env: &FMIREnvironment, expr: &MIRExpression) -> CXResult<FMIRNode> {
-    CXError::create_result(format!(
-        "FMIR conversion does not currently support expression '{}' in function '{}'",
-        expr,
-        env.current_mir_prototype().name,
-    ))
-}
-
-fn with_expression_range(mut node: FMIRNode, mir_expr: &MIRExpression) -> FMIRNode {
-    if node.source_range.is_none() {
-        node.source_range = mir_expr.source_range.as_ref().map(FMIRSourceRange::from);
-    }
-
-    node
-}
 
 pub fn convert_expression(
     env: &mut FMIREnvironment,
@@ -413,7 +20,9 @@ pub fn convert_expression(
         MIRExpressionKind::BoolLiteral(value) => Ok(FMIRNode {
             source_range: None,
             body: FMIRNodeBody::BooleanLiteral(*value),
-            _type: FMIRType::pure(MIRType::bool()),
+            _type: FMIRType::Pure {
+                mir_type: MIRType::bool(),
+            },
         }),
 
         MIRExpressionKind::IntLiteral(value, itype, signed) => Ok(FMIRNode {
@@ -648,9 +257,10 @@ pub fn convert_expression(
 
         MIRExpressionKind::Variable(name) | MIRExpressionKind::ContractVariable { name, .. } => {
             if !mir_expr._type.is_memory_reference()
-                && let Some(known) = env.query_known_value(name) {
-                    return Ok(known);
-                }
+                && let Some(known) = env.query_known_value(name)
+            {
+                return Ok(known);
+            }
 
             Ok(FMIRNode {
                 source_range: None,
@@ -731,21 +341,19 @@ pub fn convert_expression(
         MIRExpressionKind::MemoryRead { source } => {
             if let MIRExpressionKind::Variable(name)
             | MIRExpressionKind::ContractVariable { name, .. } = &source.kind
-                && let Some(mut known) = env.query_known_value(name) {
-                    if let Some(location) = env.query_memory_location(name) {
-                        let access = FMIRType::access(
-                            known._type.inner_type().clone(),
-                            vec![location],
-                            vec![],
-                        );
-                        known._type = known
-                            ._type
-                            .union(&access)
-                            .apply(known._type.inner_type().clone());
-                    }
-
-                    return Ok(known);
+                && let Some(mut known) = env.query_known_value(name)
+            {
+                if let Some(location) = env.query_memory_location(name) {
+                    let access =
+                        FMIRType::access(known._type.inner_type().clone(), vec![location], vec![]);
+                    known._type = known
+                        ._type
+                        .union(&access)
+                        .apply(known._type.inner_type().clone());
                 }
+
+                return Ok(known);
+            }
 
             let source_node = convert_expression(env, source)?;
             Ok(load_node(
