+++ compiler/cx-safe-ir/src/ast.rs
@@ -4,6 +4,8 @@ use cx_mir::mir::expression::MIRSourceRange;
 use cx_mir::mir::types::{MIRFunctionPrototype, MIRType};
 use cx_util::identifier::CXIdent;
 
+use crate::intrinsic::FMIRIntrinsicFunction;
+
 pub type FRc<T> = Rc<T>;
 
 /// Stack variable reference with depth for shadowing support
@@ -117,129 +119,6 @@ impl From<&MIRSourceRange> for FMIRSourceRange {
             end_token: value.end_token,
         }
     }
-}
-
-#[derive(Clone, Debug)]
-pub struct FMIRIntrinsicFunction {
-    pub kind: FMIRIntrinsicKind,
-}
-
-#[derive(Clone, Debug)]
-pub enum FMIRIntrinsicKind {
-    Unary(FMIRUnaryIntrinsic),
-    Binary(FMIRBinaryIntrinsic),
-    Cast(FMIRCastIntrinsic),
-}
-
-#[derive(Clone, Debug)]
-pub enum FMIRUnaryIntrinsic {
-    Neg,
-    INeg,
-    FNeg,
-    BNot,
-    LNot,
-}
-
-#[derive(Clone, Debug)]
-pub enum FMIRBinaryIntrinsic {
-    Integer {
-        bits: usize,
-        op: FMIRIntegerBinaryIntrinsicOp,
-    },
-    Float {
-        bits: usize,
-        op: FMIRFloatBinaryIntrinsicOp,
-    },
-    Pointer {
-        op: FMIRPointerBinaryIntrinsicOp,
-    },
-    PointerDiff {
-        op: FMIRPointerDiffBinaryIntrinsicOp,
-    },
-}
-
-#[derive(Clone, Debug)]
-pub enum FMIRIntegerBinaryIntrinsicOp {
-    Add,
-    Sub,
-    Mul,
-    Div,
-    Mod,
-    IMul,
-    IDiv,
-    IMod,
-    Eq,
-    Ne,
-    Lt,
-    Le,
-    Gt,
-    Ge,
-    ILt,
-    ILe,
-    IGt,
-    IGe,
-    LAnd,
-    LOr,
-    BAnd,
-    BOr,
-    BXor,
-}
-
-#[derive(Clone, Debug)]
-pub enum FMIRFloatBinaryIntrinsicOp {
-    Add,
-    Sub,
-    Mul,
-    Div,
-    Eq,
-    Ne,
-    Lt,
-    Le,
-    Gt,
-    Ge,
-}
-
-#[derive(Clone, Debug)]
-pub enum FMIRPointerBinaryIntrinsicOp {
-    Eq,
-    Ne,
-    Lt,
-    Gt,
-    Le,
-    Ge,
-}
-
-#[derive(Clone, Debug)]
-pub enum FMIRPointerDiffBinaryIntrinsicOp {
-    Add,
-    Sub,
-}
-
-#[derive(Clone, Debug)]
-pub enum FMIRCastIntrinsic {
-    Integral {
-        sextend: bool,
-        to_bits: usize,
-    },
-    FloatCast {
-        to_bits: usize,
-    },
-    PtrToInt {
-        to_bits: usize,
-    },
-    IntToPtr {
-        sextend: bool,
-    },
-    IntToFloat {
-        to_bits: usize,
-        sextend: bool,
-    },
-    FloatToInt {
-        to_bits: usize,
-        sextend: bool,
-    },
-    IntToBool,
-    ReinterpretBits,
 }
 
 #[derive(Clone, Debug)]
