+++ compiler/cx-safe-ir/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod ast;
 pub mod format;
+pub mod intrinsic;
